+++ test/specs/options/strict.spec.js
@@ -35,7 +35,7 @@ describe('strict', () => {
   it('should be ignored when the `width` option is set and its value is greater than the natural width of the image', (done) => {
     window.loadImageAsBlob('/base/docs/images/picture.jpg', (image) => {
       new Compressor(image, {
-        width: 1000,
+        width: 501,
         success(result) {
           expect(result).to.not.equal(image);
           done();
@@ -47,7 +47,7 @@ describe('strict', () => {
   it('should be ignored when the `height` option is set and its value is greater than the natural height of the image', (done) => {
     window.loadImageAsBlob('/base/docs/images/picture.jpg', (image) => {
       new Compressor(image, {
-        height: 1000,
+        height: 601,
         success(result) {
           expect(result).to.not.equal(image);
           done();
@@ -59,7 +59,7 @@ describe('strict', () => {
   it('should be ignored when the `minWidth` option is set and its value is greater than the natural width of the image', (done) => {
     window.loadImageAsBlob('/base/docs/images/picture.jpg', (image) => {
       new Compressor(image, {
-        minWidth: 1000,
+        minWidth: 501,
         success(result) {
           expect(result).to.not.equal(image);
           done();
@@ -71,7 +71,31 @@ describe('strict', () => {
   it('should be ignored when the `minHeight` option is set and its value is greater than the natural height of the image', (done) => {
     window.loadImageAsBlob('/base/docs/images/picture.jpg', (image) => {
       new Compressor(image, {
-        minHeight: 1000,
+        minHeight: 601,
+        success(result) {
+          expect(result).to.not.equal(image);
+          done();
+        },
+      });
+    });
+  });
+
+  it('should be ignored when the `maxWidth` option is set and its value is less than the natural width of the image', (done) => {
+    window.loadImageAsBlob('/base/docs/images/picture.jpg', (image) => {
+      new Compressor(image, {
+        maxWidth: 499,
+        success(result) {
+          expect(result).to.not.equal(image);
+          done();
+        },
+      });
+    });
+  });
+
+  it('should be ignored when the `maxHeight` option is set and its value is less than the natural height of the image', (done) => {
+    window.loadImageAsBlob('/base/docs/images/picture.jpg', (image) => {
+      new Compressor(image, {
+        maxHeight: 599,
         success(result) {
           expect(result).to.not.equal(image);
           done();